Move logic to separate class / refactoring
