+++ DevelopGitHub/GitHubExtension/GitHubExtension.LocalizationTool/App.xaml.cs
@@ -1,10 +1,4 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Configuration;
-using System.Data;
-using System.Linq;
-using System.Threading.Tasks;
-using System.Windows;
+﻿using System.Windows;
 
 namespace GitHubExtension.LocalizationTool
 {
