+++ DevelopGitHub/GitHubExtension/GitHubExtension.LocalizationTool/GitHubExtension.LocalizationTool.csproj
@@ -61,9 +61,11 @@
       <SubType>Designer</SubType>
     </ApplicationDefinition>
     <Compile Include="HotKeyCommands.cs" />
+    <Compile Include="Translate\JsonHelper.cs" />
     <Compile Include="Translate\Language.cs" />
     <Compile Include="Translate\NotifyProperyChangedBase.cs" />
     <Compile Include="Translate\TranslationDataRow.cs" />
+    <Compile Include="Translate\Translator.cs" />
     <Page Include="MainWindow.xaml">
       <Generator>MSBuild:Compile</Generator>
       <SubType>Designer</SubType>
