+++ DevelopGitHub/GitHubExtension/GitHubExtension.LocalizationTool/Translate/TranslationDataRow.cs
@@ -1,4 +1,4 @@
-﻿namespace GitHubExtension.LocalizationTool
+﻿namespace GitHubExtension.LocalizationTool.Translate
 {
     public class TranslationDataRow : NotifyProperyChangedBase
     {
